fix: preview
