+++ .dumi/theme/builtins/iframe.html
@@ -27,6 +27,7 @@
   <script>
     const searchParams = new URL(location.href).searchParams;
     const theme = searchParams.get('less-theme');
+    const page = searchParams.get('page');
 
     function getDist() {
       const server = searchParams.get('compilerServer');
@@ -41,31 +42,53 @@
     }
 
     function getSourceCode() {
-      const folder = searchParams.get('folder');
-      return fetch(`/sourceCode/${folder}`).then(res => res.json());
+      return fetch(`/sourceCode/${page}`).then(res => res.json());
+    }
+
+    function getExtName(name) {
+      const arr = name.split('.');
+      return `.${arr[arr.length - 1]}`;
     }
 
     function sortMiniPageFile(files) {
-      const xmlType = [];
-      const jsType = [];
-      const cssType = [];
-      const jsonType = [];
-      const restFiles = [];
-
-      files.forEach(file => {
-        if (/pages\/[a-zA-Z]*\/index.(axml|xml|html)$/.test(file)) {
-          xmlType.push(file);
-        } else if (/pages\/[a-zA-Z]*\/index.(js|ts|sjs)$/.test(file)) {
-          jsType.push(file);
-        } else if (/pages\/[a-zA-Z]*\/index.(acss|css|less|scss|sass)$/.test(file)) {
-          cssType.push(file);
-        } else if (/pages\/[a-zA-Z]*\/index.json$/.test(file)) {
-          jsonType.push(file);
-        } else {
-          restFiles.push(file);
+      const map = {
+        '.axml': 0,
+        '.js': 1,
+        '.acss': 2,
+        '.json': 3,
+      };
+      files.sort((a, b) => {
+        const arrA = a.split('/');
+        const arrB = b.split('/');
+        const lenA = arrA.length;
+        const lenB = arrB.length;
+        if (lenA === lenB) {
+          const nameA = arrA.slice(0, -1).join('/');
+          const nameB = arrB.slice(0, -1).join('/');
+          if (nameA === nameB) {
+            const result = map[getExtName(a)] - map[getExtName(b)];
+            if (result === 0) {
+              if (arrA[arrA.length - 1].startsWith('index')) {
+                return -1;
+              }
+              if (arrB[arrB.length - 1].startsWith('index')) {
+                return 1;
+              }
+              if (arrA[arrA.length - 1] < arrB[arrB.length - 1]) {
+                return -1;
+              }
+              return 1;
+            }
+            return result;
+          }
+          return nameA < nameB ? -1 : 1;
         }
+        if (lenA < lenB) {
+          return -1;
+        }
+        return 1;
       });
-      return [...xmlType, ...jsType, ...cssType, ...jsonType, ...restFiles];
+      return files;
     }
 
     async function startWebOpenBoxPreviewApp() {
@@ -81,7 +104,16 @@
       } else {
         const obj = await fetch(theme === 'dark' ? '/preview-theme-dark.json' : '/preview.json').then(res => res.json());
         dist = obj.dist;
-        sourceCode = obj.sourceCode;
+        sourceCode = {};
+        debugger
+        const str = page.replace('index', '');
+        Object.keys(obj.sourceCode).forEach(item => {
+          debugger
+          if (item.startsWith(str)) {
+            sourceCode[item] = obj.sourceCode[item];
+          }
+        });
+        debugger
       }
       sourceCode['mini.project.json'] = JSON.stringify({ 
         enableAppxNg: true,
